+++ src/components/ComponentException.php
@@ -1,0 +1,11 @@
+<?php
+namespace VovanVE\HtmlTemplate\components;
+
+use VovanVE\HtmlTemplate\Exception;
+
+/**
+ * @since 0.3.0
+ */
+class ComponentException extends Exception
+{
+}